another change

+++ build.cake
@@ -1,6 +1,29 @@
-#addin nuget:?package=Cake.Figlet&version=1.3.1
-#addin "Cake.DocFx"
-#tool "docfx.console"
+// #addin "nuget:?package=Cake.Git&version=0.15.0";
+// #addin "nuget:?package=Cake.GitPackager&version=0.1.1";
+// #addin "nuget:?package=Cake.DocFx&version=0.5.0";
+// #addin "nuget:?package=Cake.FileHelpers&version=1.0.4";
+// #tool "nuget:?package=docfx.console&version=2.26.3";
+#addin "Cake.Figlet";
+#addin "Cake.Git";
+#addin "Cake.GitPackager";
+#addin "Cake.DocFx";
+#addin "Cake.FileHelpers";
+#tool "docfx.console";
+
+
+// Helper method for setting a lot of file attributes at once
+public FilePath[] SetFileAttributes(FilePathCollection files, System.IO.FileAttributes fileAttributes)
+{
+    var results = new System.Collections.Concurrent.ConcurrentBag<FilePath>();
+
+    Parallel.ForEach(files, f =>
+    {
+        System.IO.File.SetAttributes(f.FullPath, fileAttributes);
+        results.Add(f);
+    });
+
+    return results.ToArray();
+}
 
 // Default MSBuild configuration arguments
 var configuration = Argument("configuration", "Release");
@@ -132,7 +155,10 @@ Task("CopyDocsToVersionedDirectories")
 .IsDependentOn("CloneExistingDocs")
 .Does(() =>
 {
-    DeleteDirectory("./docs/temp/latest", recursive: true);
+    if (DirectoryExists("./docs/temp/latest"))
+    {
+        DeleteDirectory("./docs/temp/latest", recursive: true);
+    }
     Information("Copying docs to docs/temp/latest");
     CopyDirectory("./docs/_site/", "./docs/temp/latest/");
